refactor/merge BaseSkill and OVOSSkill
no functional changes!

BaseSkill class removed, backwards compat import for OVOSSkill

replace duplicate private methods from BaseSkill with the public methods from OVOSSkill

reorder methods inside OVOSSkill by functionality, with comments for readability

- skill developer abstract methods / devs are meant to override these
- skill class related properties
- safe skill_id/bus wrapper properties
- magic properties -> depend on message.context / Session
- resource file loading methods
- skill object setup methods
- skill internal events and lifecycle
- intents / resource files management
- event/intent registering internal handlers
- skill developer facing utils
- event related skill developer facing utils
- intent/context skill developer facing utils
- killable_events support
- deprecated and marked for removal
