+++ ovos_workshop/skills/__init__.py
@@ -1,3 +1,5 @@
-from ovos_workshop.skills.ovos import MycroftSkill, OVOSSkill, OVOSFallbackSkill
-from ovos_workshop.skills.idle_display_skill import IdleDisplaySkill
 from ovos_workshop.decorators.layers import IntentLayers
+from ovos_workshop.skills.ovos import OVOSSkill, OVOSFallbackSkill
+from ovos_workshop.skills.idle_display_skill import IdleDisplaySkill
+from ovos_workshop.skills.mycroft_skill import MycroftSkill
+
